% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quantgen_trials}
\alias{quantgen_trials}
\title{Multiple trials for a given set of parameter values.}
\usage{
quantgen_trials(n_trials, N0, P0, V0, U0, r, a, f, b, c, m, g, etaN, etaP,
  sig2N, sig2P, delta, delta2, start_t, max_t, density_threshold = 1e-05,
  precision = 0.01, n_cores = 1L, show_progress = TRUE,
  return_every = 0L)
}
\arguments{
\item{n_trials}{Number of trials to perform.}

\item{N0}{Vector containing the starting abundance(s) for each resource.}

\item{P0}{Vector containing the starting abundance(s) for each consumer.}

\item{V0}{Matrix containing the starting trait value(s) for each resource.
Each row refers to a particular resource.}

\item{U0}{Matrix containing the starting trait value(s) for each consumer.
Each row refers to a particular consumer.}

\item{r}{Base growth rate for resources.}

\item{a}{Baseline competition coefficient for resources.}

\item{f}{Cost of defensive traits for resources.}

\item{b}{Coefficient for attack rate of consumers on resources.}

\item{c}{Base growth rate for consumers.}

\item{m}{Mortality rate for consumers.}

\item{g}{Cost of offensive traits for consumers.}

\item{etaN}{Non-additive effect of increasing 2+ traits for resources.}

\item{etaP}{Non-additive effect of increasing 2+ traits for consumers.}

\item{sig2N}{Additive genetic variance for resources.}

\item{sig2P}{Additive genetic variance for consumers.}

\item{delta}{SD for log-normal distribution generating perturbation after the
starting time steps.}

\item{delta2}{SD for log-normal distribution generating perturbation at the
halfway point in the primary time steps.}

\item{start_t}{Number of starting time steps.}

\item{max_t}{Number of primary time steps.}

\item{density_threshold}{Threshold for extinction. Defaults to \code{1e-5}.}

\item{precision}{Threshold for considering rows of traits to be the same.
Defaults to \code{1e-2}.}

\item{n_cores}{Number of cores to use for processing.
Ignored if not compiled using OpenMP. Defaults to \code{1}.}

\item{show_progress}{Boolean for whether to show a progress bar. Defaults to \code{TRUE}.}

\item{return_every}{Integer specifing the number of iterations (after the starting
iterations) at which to output values of abundances and trait values.
Values <= 0 mean that only final values are output.
NOTE: A high value of \code{max_t} and a low value of \code{return_every} will result in
a massive object being output.
Defaults to \code{0}.}
}
\value{
A list containing the following components:
\itemize{
\item \code{resource}: Info for resource(s).
\item \code{consumer}: Info for consumer(s).
}

Both \code{resource} and \code{consumer} have the same structure:
\itemize{
\item If \code{return_every <= 0}, a matrix with the following columns (in order):
\itemize{
\item The trial number
\item The final fitness
\item The final selection pressure
\item The final trait values (1 column for each trait)
}
\item If \code{return_every > 0}, a list with the following items:
\itemize{
\item \code{final}: Matrix of final values for each trial (just like if
\code{return_every <= 0}).
\item \code{past_abund}: A list of matrices (one for each trial).
Each matrix contains species abundances through time,
where rows are species and columns are time.
\item \code{past_traits}: A list of 3D arrays (one for each rial).
Each array contains species traits through time,
where rows are species, columns are traits, "slices" are time.
}
}
}
\description{
Multiple trials for a given set of parameter values.
}
