% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgenCR.R
\docType{package}
\name{quantgenCR}
\alias{quantgenCR}
\alias{quantgenCR-package}
\title{quantgenCR: Quantitative genetics for a consumer-resource model.}
\description{
quantgenCR: Quantitative genetics for a consumer-resource model.
}
